CovidAnalyzerApi/requirements.txt
ansicon==1.89.0
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-heroku==0.3.1
django-picklefield==3.0.1
django-q==1.3.4
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jinxed==1.0.1
joblib==1.0.0
jsonpickle==1.5.0
MarkupSafe==1.1.1
nltk==3.5
openapi-codec==1.3.2
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
searchtweets-v2==1.0.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tagme==0.1.3
textblob==0.15.3
tqdm==4.55.2
twitter-sentiment==0.0.6.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
whitenoise==5.2.0
